+++ src/sidplayfp/sidplayfp.h
@@ -118,7 +118,7 @@ public:
      * and #isPlaying() is true an error occurred, use #error() to get
      * a detailed message.
      */
-    uint_least32_t play(short *buffer, uint_least32_t count);
+    uint_least32_t play(float *buffer, uint_least32_t count);
 
     /**
      * Check if the engine is playing or stopped.
